+++ tools/Gruntfile.js
@@ -25,12 +25,25 @@ module.exports = function(grunt) {
         //                dest: 'dist/FILE_NAME.min.js'
         //            }
         //        },
+        compileAppSetupToStartupSequence: {
+            files: ['../tests/minifierFullMapAppSetup.json']
+        },
         watch: {
-            //            files: '<config:lint.files>',
-            files: ['../applications/**/*.js', '../bundles/**/*.js', '../libraries/**/*.js', '../packages/**/*.js', '../resources/**/*.js', '../sources/**/*.js', '../tests/**/*.js'],
-            // uncommented as validate causes unnecessary delay
-            //            tasks: ['validate', 'compile', 'testacularRun:dev', 'yuidoc:dist']
-            tasks: ['compileDev', 'testacularRun:dev']
+            appsetup: {
+                files: '<%= compileAppSetupToStartupSequence.files %>',
+                tasks: ['compileAppSetupToStartupSequence']
+            },
+            src: {
+                //            files: '<config:lint.files>',
+                files: ['../applications/**/*.js', '../bundles/**/*.js', '../libraries/**/*.js', '../packages/**/*.js', '../resources/**/*.js', '../sources/**/*.js'],
+                // uncommented as validate causes unnecessary delay
+                //            tasks: ['validate', 'compile', 'testacularRun:dev', 'yuidoc:dist']
+                tasks: ['compileDev', 'testacularRun:dev']
+            },
+            test: {
+                files: ['../tests/**/*.js'],
+                tasks: ['testacularRun:dev']
+            }
         },
         sprite: {
             options: {
@@ -107,7 +120,48 @@ module.exports = function(grunt) {
         grunt.config.set("compile.dev.options", this.options());
 
         grunt.task.run('compile');
-    })
+    });
+
+    grunt.registerTask('compileAppSetupToStartupSequence', function() {
+        var done = this.async();
+        var starttime = (new Date()).getTime();
+
+        grunt.log.writeln('Running compile AppSetup to startupSequence...');
+
+        console.log('check', grunt.config('watch').appsetup);
+
+        // read files and parse output name
+        var files = grunt.config(this.name).files[0],
+            outputFilename = files.replace(".json", ".opts.js");
+
+        // read file
+        var fs = require('fs'),
+            cfgFile = fs.readFileSync(files, 'utf8');
+
+        // convert to usable format
+        var startupSequence = JSON.parse(cfgFile).startupSequence,
+            definedBundles = {},
+            bundle,
+            result = "var _defaultsStartupSeq = ";
+
+        // loop startup sequence bundles and add to hashmap of defined bundles
+        for(var i = 0, ilen = startupSequence.length; i < ilen; i++) {
+            bundle = startupSequence[i];
+            // Add here code to change bundlePath to "ignored/butRequiredToBeInMinifierFullMapAppSetupInTests/packages/framework/bundle/"
+            // or something similar as the bundlePaths are not used with minifierAppSetup
+            definedBundles[bundle.bundlename] = bundle;
+        }
+
+        // add stringified bundle definitions
+        result +=JSON.stringify(definedBundles);
+
+        // write file to be used in testing as is
+        fs.writeFileSync(outputFilename, result , 'utf8');
+
+        var endtime = (new Date()).getTime();
+        grunt.log.writeln('compileAppSetupToStartupSequence completed in ' + ((endtime - starttime) / 1000) + ' seconds');
+        done();
+    });
 
     grunt.registerTask('release', 'Release build', function(version, configs) {
         var apps = [],
@@ -133,12 +187,31 @@ module.exports = function(grunt) {
                     resultCSSName: "../dist/<%= version %>" + appName + "/css/icons.css",
                     spritePathInCSS: "../icons"
                 },
-                files = [{
+                files = [],
+                copyFiles = {
                     "expand": true,
                     "cwd": cwd + "/",
                     "src": ["css/**", "images/**", "*.js"],
                     "dest": dest
-                }];
+                };
+
+            // subsets have underscore (_) in appName, which means we need to
+            // get the parent resources first and then replace with subset specific stuff
+            var parentAppName = appName.substring(0, appName.indexOf('_'));
+            if (appName !== parentAppName) {
+                // copy files from parent folder to be replaced by child
+                files.push({
+                    "expand": true,
+                    "cwd": cwd.replace(appName, parentAppName) + "/",
+                    "src": ["css/**", "images/**", "*.js"],
+                    "dest": dest
+                });
+                // modify css-sprite to use parent icons instead
+                options.iconDirectoryPath = options.iconDirectoryPath.replace(appName, parentAppName);
+            }
+
+            // add files to be copied
+            files.push(copyFiles);
 
             // setting task configs
             grunt.config.set("copy." + appName + ".files", files);
@@ -158,7 +231,7 @@ module.exports = function(grunt) {
         grunt.task.run('compile');
         grunt.task.run('sprite');
         grunt.task.run('yuidoc');
-        //        grunt.task.run('mddocs');
+        //grunt.task.run('mddocs');
     });
 
     grunt.registerTask('packageopenlayer', 'Package openlayers according to packages', function(packages) {