+++ ReleaseNotes.md
@@ -1,5 +1,15 @@
 # Release Notes
 
+## 1.9
+
+### toolbar bundle
+
+Added a way to disable a button by default from configuration.
+
+### promote bundle
+
+Promote login and registering by replacing the real bundle for guest users. Configurable tile, flyout and toolbar buttons.
+
 ## 1.8
 
 ### sandbox/map-layer-service